Corrects datatype of operation sequence

+++ rest/1.0/cluster/admin/v1/pdu/cluster_admin_request.go
@@ -11,5 +11,5 @@ type ClusterRetrieveRequest struct {
 
 type ClusterOperationRequest struct {
 	ClusterRetrieveRequest
-	OperationSeq int32 `json:"operation_seq"`
+	OperationSeq uint64 `json:"operation_seq"`
 }
